Move code to `test.ts`
